+++ backend/app/tasks/import/apiFootballData/matches.job.ts
@@ -1,8 +1,8 @@
 import { from, lastValueFrom } from 'rxjs';
 import { mergeMap } from 'rxjs/operators';
 
-import { MatchRepository } from '../../db/repositories/match.repo.js';
-import { FootballApiClient } from '../../thirdParty/footballApi/apiClient.js';
+import { MatchRepository } from '../../../../db/repositories/match.repo.js';
+import { FootballApiClient } from '../../../../thirdParty/footballApi/apiClient.js';
 import { Job } from '../jobs/job.js';
 import { Queue } from '../queue.js';
 import Builder from './matchesJob.builder.js';
